requirements.txt
django-registration-redux
Django==1.7
django-crispy-forms==1.4.0
psycopg2==2.5.4
django-email-changer==0.1.2
